my cache support timeout

+++ include/cavan/cache.h
@@ -70,5 +70,5 @@ void cavan_cache_open(struct cavan_cache *cache);
 void cavan_cache_close(struct cavan_cache *cache);
 char *cavan_cache_pointer_add(struct cavan_cache *cache, char *pointer, off_t offset);
 ssize_t cavan_cache_write(struct cavan_cache *cache, const char *buff, size_t size);
-ssize_t cavan_cache_read(struct cavan_cache *cache, char *buff, size_t size, size_t reserved, long timeout);
-ssize_t cavan_cache_fill(struct cavan_cache *cache, char *buff, size_t size, size_t reserved, long timeout);
+ssize_t cavan_cache_read(struct cavan_cache *cache, char *buff, size_t size, size_t reserved, u32 timeout);
+ssize_t cavan_cache_fill(struct cavan_cache *cache, char *buff, size_t size, size_t reserved, u32 timeout);
